+++ src/cache/test/getRestoreCacheCommand.test.ts
@@ -1,5 +1,5 @@
 import getRestoreCacheCommand from 'cache/getRestoreCacheCommand';
-import normalize from 'test-tool/helpers/normalize';
+import normalize from 'test-tool/helper/normalize';
 
 describe('getRestoreCacheCommand', () => {
   test('simple key property', () => {
