+++ src/config/test/AllConfigFiles.spec.ts
@@ -1,8 +1,8 @@
 import Types from 'common/Types';
 import AllConfigFiles from 'config/AllConfigFiles';
-import AppTestHarness from 'test-tool/helpers/AppTestHarness';
+import AppTestHarness from 'test-tool/helper/AppTestHarness';
 import FakeEditorGateway from 'gateway/FakeEditorGateway';
-import getContextStub from 'test-tool/helpers/getContextStub';
+import getContextStub from 'test-tool/helper/getContextStub';
 
 function getMockContext(filePath: string) {
   const initialContext = getContextStub();
