+++ src/job/test/JobProvider.spec.ts
@@ -1,5 +1,5 @@
-import AppTestHarness from 'test-tool/helpers/AppTestHarness';
-import getContextStub from 'test-tool/helpers/getContextStub';
+import AppTestHarness from 'test-tool/helper/AppTestHarness';
+import getContextStub from 'test-tool/helper/getContextStub';
 import JobProviderFactory from 'job/JobProviderFactory';
 import JobFactory from 'job/JobFactory';
 import JobTreeItem from 'job/JobTreeItem';
